feat: add support for getting a single actor and a single movie
Add two new routes to the API, `/api/actors/<int:actor_id>` and `/api/movies/<int:movie_id>`, to support getting a single actor and a single movie, respectively. These routes are protected by the `get:actors` and `get:movies` permissions, respectively.

feat: add edit functionality to entity CRUD classes

Add a new `edit` method to the `EntityCRUD` class that fetches the entity data from the API and populates the form fields with the data. Also update the `display` method of the `ActorCRUD` and `MovieCRUD` classes to check if the list element exists before updating it.

refactor: rename form input ids in actor and movie templates

Rename the `actor-name` and `actor-bio` input ids in the `actor.html` template to `name` and `bio`,