feat(app.py, entity_crud.js, home.html): add JWT token authentication
Add JWT token authentication to the Flask app and the JavaScript code that fetches data from the API. The token is stored in the local storage of the browser and sent as an Authorization header in the fetch requests. The token is also added to the home page as a script tag.